// my-app/types/database.ts
export type Database = {

  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      cart_items: {
        Row: {
          cart_id: string
          created_at: string
          id: number
          product_id: number
          quantity: number
        }
        Insert: {
          cart_id: string
          created_at?: string
          id?: number
          product_id: number
          quantity?: number | 1
        }
        Update: {
          cart_id?: string
          created_at?: string
          id?: number
          product_id?: number
          quantity?: number
        }
        Relationships: [
          {
            foreignKeyName: "cart_items_cart_id_fkey"
            columns: ["cart_id"]
            isOneToOne: false
            referencedRelation: "carts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "cart_items_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "product_variants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      product_variants: {
        Row: {
          id: number
          is_default: boolean | null
          price: number
          product_id: number
          sku: string
          stock_quantity: number
          weight_unit: string
          weight_value: number
        }
        Insert: {
          id?: number
          is_default?: boolean | null
          price: number
          product_id: number
          sku: string
          stock_quantity?: number
          weight_unit: string
          weight_value: number
        }
        Update: {
          id?: number
          is_default?: boolean | null
          price?: number
          product_id?: number
          sku?: string
          stock_quantity?: number
          weight_unit?: string
          weight_value?: number
        }
        Relationships: [
          {
            foreignKeyName: "product_variants_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
        ]
      }
      products: {
        Row: {
          category_id: number
          id: number
          name: string
        }
        Insert: {
          category_id: number
          id?: number
          name: string
        }
        Update: {
          category_id?: number
          id?: number
          name?: string
        }
        Relationships: [
          {
            foreignKeyName: "products_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      order_status:
      | "pending"
      | "processing"
      | "returned"
      | "delivered"
      | "cancelled"
      | "shipped"
      roles: "owner" | "admin" | "customer"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}